Add test suite and initial support for PyPI publishing via Travis CI
This changeset adds a basic `unittest` suite for this project, along
with instructions for executing it locally and inside of a Docker
container. In addition, support for Travis CI was added to execute the
test suite after linting.

PyPI deployment has not been tested, but it will only be triggered on
Git tags.